[test] #100 Point 테스트 코드 수정
